backend/requirements.txt
# HTTP requests and HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.2

# JSON handling
ujson>=5.8.0

# File paths and system utilities (standard library)
# pathlib is built-in, no need to install

# CSV reading and logging (standard library)
# csv, logging, time are built-in

# Progress bar
tqdm>=4.65.0

# Sentence embeddings
sentence-transformers>=2.2.2
torch>=2.1.0

# Vector search
faiss-cpu>=1.7.4
numpy>=1.26.0

# Web API
fastapi>=0.111.0
pydantic>=2.5.1
uvicorn[standard]>=0.23.2

# Ollama client
ollama>=0.1.0
